Refactore Console Command Directory
