+++ Console/BlockListCommand.php
@@ -1,43 +1,43 @@
 <?php
 declare(strict_types=1);
 
-namespace Firegento\ContentProvisioning\Model\Console;
+namespace Firegento\ContentProvisioning\Console;
 
-use Firegento\ContentProvisioning\Api\Data\EntryInterface;
-use Firegento\ContentProvisioning\Api\Data\PageEntryInterface;
-use Firegento\ContentProvisioning\Model\Query\GetPageEntryList\Proxy as GetPageEntryList;
-use Firegento\ContentProvisioning\Model\Query\GetPagesByPageEntry\Proxy as GetPagesByPageEntry;
+use Firegento\ContentProvisioning\Api\Data\BlockEntryInterface;
+use Firegento\ContentProvisioning\Model\Query\GetBlockEntryList\Proxy as GetBlockEntryList;
+use Firegento\ContentProvisioning\Model\Query\GetBlocksByBlockEntry\Proxy as GetBlocksByBlockEntry;
 use Magento\Framework\Exception\LocalizedException;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Helper\Table;
 
-class PageListCommand extends Command
+class BlockListCommand extends Command
 {
     /**
-     * @var GetPageEntryList
+     * @var GetBlockEntryList
      */
-    private $getAllContentEntries;
+    private $getAllBlockEntries;
 
     /**
-     * @var GetPagesByPageEntry
+     * @var GetBlocksByBlockEntry
      */
-    private $getPagesByPageEntry;
+    private $getBlocksByBlockEntry;
 
     /**
-     * @param GetPageEntryList $getAllContentEntries
-     * @param GetPagesByPageEntry $getPagesByPageEntry
-     * @param string|null $name
+     * @param GetBlockEntryList     $getAllBlockEntries
+     * @param GetBlocksByBlockEntry $getBlocksByBlockEntry
+     * @param string|null           $name
      */
     public function __construct(
-        GetPageEntryList $getAllContentEntries,
-        GetPagesByPageEntry $getPagesByPageEntry,
-        string $name = null
+        GetBlockEntryList     $getAllBlockEntries,
+        GetBlocksByBlockEntry $getBlocksByBlockEntry,
+        string                $name = null
     ) {
         parent::__construct($name);
-        $this->getAllContentEntries = $getAllContentEntries;
-        $this->getPagesByPageEntry = $getPagesByPageEntry;
+
+        $this->getAllBlockEntries    = $getAllBlockEntries;
+        $this->getBlocksByBlockEntry = $getBlocksByBlockEntry;
     }
 
     /**
@@ -48,8 +48,7 @@ class PageListCommand extends Command
         $table = new Table($output);
         $table->setHeaders(['Key', 'Identifier', 'Stores', 'Maintained', 'Active', 'Title', 'in DB (IDs)']);
 
-        /** @var EntryInterface $entry */
-        foreach ($this->getAllContentEntries->get() as $entry) {
+        foreach ($this->getAllBlockEntries->get() as $entry) {
             $table->addRow([
                 $entry->getKey(),
                 $entry->getIdentifier(),
@@ -57,7 +56,7 @@ class PageListCommand extends Command
                 $entry->isMaintained() ? 'yes' : 'no',
                 $entry->isActive() ? 'yes' : 'no',
                 $entry->getTitle(),
-                $this->getExistsInDbValue($entry),
+                $this->getExistsInDbValue($entry)
             ]);
         }
 
@@ -69,20 +68,20 @@ class PageListCommand extends Command
      */
     protected function configure()
     {
-        $this->setName('content-provisioning:page:list');
-        $this->setDescription('List all configured CMS page entries');
+        $this->setName('content-provisioning:block:list');
+        $this->setDescription('List all configured CMS block entries');
         parent::configure();
     }
 
     /**
-     * @param PageEntryInterface $entry
+     * @param BlockEntryInterface $entry
      * @return string
      */
-    private function getExistsInDbValue(PageEntryInterface $entry): string
+    private function getExistsInDbValue(BlockEntryInterface $entry): string
     {
         try {
             $ids = [];
-            foreach ($this->getPagesByPageEntry->execute($entry) as $page) {
+            foreach ($this->getBlocksByBlockEntry->execute($entry) as $page) {
                 $ids[] = $page->getId();
             }
