+++ Console/ExportCommand.php
@@ -1,0 +1,19 @@
+<?php
+declare(strict_types=1);
+
+namespace Firegento\ContentProvisioning\Console;
+
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class ExportCommand extends Command
+{
+    /**
+     * {@inheritdoc}
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+
+    }
+}
